requirements.txt
Mako==1.0.0
MarkupSafe==0.23
PasteDeploy==1.5.2
TornadIO2==0.0.4
WebOb==1.4
anyjson==0.3.3
backports.ssl-match-hostname==3.4.0.2
certifi==14.05.14
chaussette==1.2
circus==0.11.1
circus-web==0.5
configobj==5.0.6
cornice==0.17
gevent==1.0.1
greenlet==0.4.5
iowait==0.2
kafka-python==0.9.2
psutil==2.1.3
pyramid==1.5.2
pyzmq==14.4.1
repoze.lru==0.6
simplejson==3.6.5
six==1.8.0
tomako==0.1.0
tornado==4.0.2
translationstring==1.3
venusian==1.0
zope.deprecation==4.1.1
zope.interface==4.1.1
